refactor: add id string check for positive integer value

+++ src/modules/blog-content/infrastructure/sql/blogs.repository.ts
@@ -3,12 +3,17 @@ import { BlogType, CreateBlogRepoParams, UpdateBlogRepoParams } from '../../type
 import { Pool } from 'pg';
 import { PG_POOL } from '../../../../common/constants.js';
 import { BlogNotFoundDomainException } from '../../../../common/exceptions/domain-exceptions.js';
+import { isPositiveIntegerString } from '../../../../common/helpers/is-positive-integer-string.js';
 
 @Injectable()
 export class BlogsRepository {
   constructor(@Inject(PG_POOL) private readonly pool: Pool) {}
 
   async getBlog(id: string): Promise<BlogType> {
+    if (!isPositiveIntegerString(id)) {
+      throw new BlogNotFoundDomainException();
+    }
+
     const result = await this.pool.query(
       `
         SELECT *
@@ -35,6 +40,10 @@ export class BlogsRepository {
   }
 
   async checkBlogExists(id: string): Promise<void> {
+    if (!isPositiveIntegerString(id)) {
+      throw new BlogNotFoundDomainException();
+    }
+
     const result = await this.pool.query(
       `
         SELECT EXISTS(SELECT 1 FROM blogs WHERE id = $1::int) AS exists
@@ -66,6 +75,11 @@ export class BlogsRepository {
   }
   async updateBlog(params: UpdateBlogRepoParams): Promise<void> {
     const { id, name, description, websiteUrl } = params;
+
+    if (!isPositiveIntegerString(id)) {
+      throw new BlogNotFoundDomainException();
+    }
+
     const result = await this.pool.query(
       `
         UPDATE blogs
@@ -81,6 +95,10 @@ export class BlogsRepository {
   }
 
   async deleteBlog(id: string): Promise<void> {
+    if (!isPositiveIntegerString(id)) {
+      throw new BlogNotFoundDomainException();
+    }
+
     const result = await this.pool.query(
       `
         DELETE FROM blogs
